reading user name from firestore and use it in profil view
