+++ src/index.html
@@ -13,9 +13,9 @@
   <section id="main">
     
   </section>
-
+  
   <script src="https://www.gstatic.com/firebasejs/7.14.0/firebase-app.js"></script>
-  <script>
+  <!--<script>
     
     var firebaseConfig = {
       apiKey: "AIzaSyDHkDChMugsG2GomKvzryoKiEcnVk-IPPk",
@@ -28,7 +28,22 @@
     };
     
     firebase.initializeApp(firebaseConfig);
-  </script> 
+  </script> -->
+
+  <script>
+    // Your web app's Firebase configuration
+    var firebaseConfig = {
+      apiKey: "AIzaSyCWtB_El7FwC2HVhoL2kfuFoqc12okRJd4",
+      authDomain: "usuarios-triplife.firebaseapp.com",
+      databaseURL: "https://usuarios-triplife.firebaseio.com",
+      projectId: "usuarios-triplife",
+      storageBucket: "usuarios-triplife.appspot.com",
+      messagingSenderId: "369850574673",
+      appId: "1:369850574673:web:2c3ed8255fa9182d129570"
+    };
+    // Initialize Firebase
+    firebase.initializeApp(firebaseConfig);
+  </script>
 
   <script src="https://www.gstatic.com/firebasejs/7.14.0/firebase-auth.js"></script>
   <script src="https://www.gstatic.com/firebasejs/7.14.0/firebase-firestore.js"></script>
