+++ src/user.js
@@ -1,16 +1,20 @@
 export default class User {
     
+    /*
     name
     lastName
     email
     photo
+    password
     description
-    uid    
+    date
+    uid*/    
 
-    constructor(name,lastName,email,photo,description,uid){
+    constructor(name,lastName,email,password,photo,description,date,uid){
         this.name = name;
         this.lastName = lastName;
         this.email = email;
+        this.password = password;
         this.photo = photo;
         this.description = description;
         this.date = date;
